test: share the PR-monitor integration fakes across the runner part files
Each ``test_pr_monitor_runner_part_*`` module carried a byte-identical
218-line copy of ``FakeAdapter``, ``RecordedSleep``, ``_git_calls``,
``_pr_payload``, the four fixtures and ``_seed_monitoring_workspace``. Two of
the three parts sat at 1487 lines with the guardrail at 1500, so the next
regression added to either would have failed the line-limit test the same way
comment_verdict.py just did.

Move the shared block into ``tests/integration/runtime/_pr_monitor_runner_fixtures``
and load it as a pytest plugin, matching the existing
``tests.unit.runtime._verdict_retry_fixtures`` convention. ``_make_runner``
stays in each part because it has diverged between them. No test bodies
change; the parts drop to 1227/1269/1269 lines.
